bugfix: period of 500us and 250us pulses was incorrect (2x too fast)
The CKE_250US and CKE_500US pulse trains were being generated incorrectly. This produced a series of pulses which had a period of half the intended value (i.e. 125us for CKE_250US and 250us for CKE_500US). The effect of this was to completely foul up the HSTMD detector, among other things.