[index freq measurement] give index pulse priority over clock increment
Resolve potential issue where an index pulse comes in at the exact same time as an index frequency counter increment, causing a missed store. Now this behaviour is reversed: if both come in at the same time, the frequency value will be latched, and the clock will be missed. This gives a potential 1clk error for the index frequency counter (or about the same as an off-the-shelf frequency counter).